+++ internal/acctest/template/usage_example_test.go
@@ -15,7 +15,7 @@ func TestExpectedBehaviour(t *testing.T) {
 resource "aiven_kafka" "bar" {
   project                 = data.aiven_project.foo.project
   cloud_name              = "google-europe-west1"
-  plan                    = "startup-2"
+  plan                    = "startup-4"
   service_name            = "{{ .service_name }}"
   maintenance_window_dow  = "monday"
   maintenance_window_time = "10:00:00"
@@ -29,7 +29,7 @@ provider "google" {
 }`)
 
 	builder := templateSet.NewBuilder().
-		AddDataSource("aiven_project", map[string]any{ //provide configuration for the data source
+		AddDataSource("aiven_project", map[string]any{ // provide configuration for the data source
 			"resource_name": "foo",
 			"project":       "test_project",
 		}).