+++ internal/sdkprovider/service/serviceintegration/service_integration_test.go
@@ -209,7 +209,7 @@ data "aiven_project" "foo" {
 resource "aiven_kafka" "kafka1" {
   project                 = data.aiven_project.foo.project
   cloud_name              = "google-europe-west1"
-  plan                    = "startup-2"
+  plan                    = "startup-4"
   service_name            = "test-acc-sr-kafka-%s"
   maintenance_window_dow  = "monday"
   maintenance_window_time = "10:00:00"
@@ -268,7 +268,7 @@ data "aiven_project" "foo" {
 resource "aiven_kafka" "source" {
   project                 = data.aiven_project.foo.project
   cloud_name              = "google-europe-west1"
-  plan                    = "startup-2"
+  plan                    = "startup-4"
   service_name            = "test-acc-sr-source-%s"
   maintenance_window_dow  = "monday"
   maintenance_window_time = "10:00:00"
@@ -292,7 +292,7 @@ resource "aiven_kafka_topic" "source" {
 resource "aiven_kafka" "target" {
   project                 = data.aiven_project.foo.project
   cloud_name              = "google-europe-west1"
-  plan                    = "startup-2"
+  plan                    = "startup-4"
   service_name            = "test-acc-sr-target-%s"
   maintenance_window_dow  = "monday"
   maintenance_window_time = "10:00:00"
@@ -584,7 +584,7 @@ data "aiven_project" "project" {
 resource "aiven_kafka" "kafka" {
   project                 = data.aiven_project.project.project
   cloud_name              = "google-europe-west1"
-  plan                    = "startup-2"
+  plan                    = "startup-4"
   service_name            = "%[1]s-kafka"
   maintenance_window_dow  = "monday"
   maintenance_window_time = "10:00:00"